Replace Belgic Confession from sticky header to pager navigation
