+++ lib/navigations/navigation_page.dart
@@ -6,7 +6,7 @@ import '../creeds/chalcedonian_creed/chalcedoniancreed_page.dart';
 import '../creeds/nicene_creed/nicenecreed_page.dart';
 import '../councils/council_of_orange/counciloforangecanon_page.dart';
 import '../councils/second_council_of_contantionople/second_council_of_contantionople_page.dart';
-import '../confessions/belgic_confession/belgic_page.dart';
+import '../confessions/belgic_confession/belgic_confession_page.dart';
 import '../confessions/westminster_confession/westminster_confession_page.dart';
 import '../catechisms/heidelberg_catechism/heidelberg_page.dart';
 import '../other_documents/ninetyfive_theses/ninteyfivethesese_page.dart';
